Reorganize malloc code, mark #![no_std]
